chore: update package configuration and remove app.ts
- Changed the main entry point in package.json from src/app.js to dist/index.js.
- Updated development scripts to use ts-node for local development.
- Added engine requirements for Node.js version in package.json.
- Introduced new devDependencies including ts-node and updated package-lock.json.
- Removed the app.ts file as part of the refactor.
