--- nyt-express-backend/src/app.ts
@@ -1,42 +1,0 @@
-import express, { Request, Response, NextFunction } from 'express';
-import mongoose from 'mongoose';
-import nytRoutes from './routes/nytRoutes';
-import dotenv from 'dotenv';
-import cors from 'cors';
-
-dotenv.config();
-
-const app = express();
-const PORT = process.env.PORT || 3000;
-
-// Configure CORS to allow requests only from http://localhost:3000
-const corsOptions = {
-    origin: 'http://localhost:3000', // Replace with your frontend's URL
-    methods: ['GET', 'POST', 'PUT', 'DELETE'], // Allowed HTTP methods
-    allowedHeaders: ['Content-Type', 'Authorization'], // Allowed headers
-};
-
-app.use(cors(corsOptions));
-app.use(express.json());
-
-mongoose.connect(process.env.DATABASE_URL as string)
-    .then(() => {
-        console.log('MongoDB connected');
-        app.listen(PORT, () => {
-            console.log(`Server is running on port ${PORT}`);
-        });
-    })
-    .catch(err => {
-        console.error('Database connection error:', err);
-    });
-
-app.use('/api', nytRoutes);
-
-
-// Error handling middleware
-app.use((err: Error, req: Request, res: Response, next: NextFunction) => {
-    console.error(err.stack);
-    res.status(500).send('Something broke!');
-});
-
-export default app;
